serving/requirements.txt
flask
pandas==1.0.5
numpy==1.19.1
torch
requests
keras
google-cloud